+++ src/settings-ui/Settings.UI.UnitTests/Cmd/SetSettingCommandTests.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
+using System.Collections.Generic;
 using System.IO.Abstractions.TestingHelpers;
 
 using Microsoft.PowerToys.Settings.UI.Library;
@@ -39,27 +40,62 @@ public class SetSettingCommandTests
 
     // Each setting has a different type.
     [TestMethod]
-    [DataRow(typeof(PowerRenameSettings), nameof(PowerRenameProperties.MaxMRUSize), "123")]
-    [DataRow(typeof(FancyZonesSettings), nameof(FZConfigProperties.FancyzonesBorderColor), "#00FF00")]
-    [DataRow(typeof(MeasureToolSettings), nameof(MeasureToolProperties.ActivationShortcut), "Ctrl+Alt+Delete")]
-    [DataRow(typeof(AlwaysOnTopSettings), nameof(AlwaysOnTopProperties.SoundEnabled), "False")]
-    [DataRow(typeof(AlwaysOnTopSettings), nameof(AlwaysOnTopProperties.ShowInSystemMenu), "true")]
-    [DataRow(typeof(PowerAccentSettings), nameof(PowerAccentProperties.ShowUnicodeDescription), "true")]
     [DataRow(typeof(AwakeSettings), nameof(AwakeProperties.Mode), "EXPIRABLE")]
     [DataRow(typeof(AwakeSettings), nameof(AwakeProperties.ExpirationDateTime), "March 31, 2020 15:00 +00:00")]
-    [DataRow(typeof(PowerLauncherSettings), nameof(PowerLauncherProperties.MaximumNumberOfResults), "322")]
-
-    [DataRow(typeof(ColorPickerSettings), nameof(ColorPickerProperties.CopiedColorRepresentation), "RGB")]
+    [DataRow(typeof(LightSwitchSettings), nameof(LightSwitchProperties.LightTime), "600")]
+    [DataRow(typeof(LightSwitchSettings), nameof(LightSwitchProperties.EnableDarkModeProfile), "true")]
+    [DataRow(typeof(MonitorSettings), nameof(MonitorProperties.ScanIntervalSeconds), "3600")]
+    [DataRow(typeof(PowerDisplaySettings), nameof(PowerDisplayProperties.MonitorRefreshDelay), "10")]
     public void SetModuleSetting(Type moduleSettingsType, string settingName, string newValueStr)
     {
         SetSetting(moduleSettingsType, settingName, newValueStr);
     }
 
-    [DataRow(typeof(GeneralSettings), "Enabled.MouseWithoutBorders", "true")]
+    [DataRow(typeof(GeneralSettings), "Enabled.Monitor", "true")]
     [DataRow(typeof(GeneralSettings), nameof(GeneralSettings.AutoDownloadUpdates), "true")]
     [TestMethod]
     public void SetGeneralSetting(Type moduleSettingsType, string settingName, string newValueStr)
     {
         SetSetting(moduleSettingsType, settingName, newValueStr);
+    }
+
+    [TestMethod]
+    [DataRow("FancyZones.FancyzonesBorderColor", "#00FF00")]
+    [DataRow("PowerLauncher.MaximumNumberOfResults", "322")]
+    [DataRow("GeneralSettings.Enabled.MouseWithoutBorders", "true")]
+    public void SetSettingShouldRejectInactiveModuleSurface(string settingName, string settingValue)
+    {
+        Assert.ThrowsException<ArgumentException>(() => Execute(settingName, settingValue, settingsUtils));
+    }
+
+    [TestMethod]
+    public void GetSettingShouldAcceptGeneralAliasAndActiveModuleSurface()
+    {
+        var requestedSettings = new Dictionary<string, List<string>>
+        {
+            ["General"] = [nameof(GeneralSettings.AutoDownloadUpdates)],
+            [MonitorSettings.ModuleName] = [nameof(MonitorProperties.ScanIntervalSeconds)],
+        };
+
+        var result = GetSettingCommandLineCommand.Execute(requestedSettings);
+
+        StringAssert.Contains(result, "\"General\"");
+        StringAssert.Contains(result, $"\"{MonitorSettings.ModuleName}\"");
+        StringAssert.Contains(result, nameof(GeneralSettings.AutoDownloadUpdates));
+        StringAssert.Contains(result, nameof(MonitorProperties.ScanIntervalSeconds));
+    }
+
+    [TestMethod]
+    [DataRow("FancyZones", nameof(FZConfigProperties.FancyzonesBorderColor))]
+    [DataRow("MouseWithoutBorders", "UseService")]
+    [DataRow("BogusModule", "Setting")]
+    public void GetSettingShouldRejectInactiveModuleSurface(string moduleName, string settingName)
+    {
+        var requestedSettings = new Dictionary<string, List<string>>
+        {
+            [moduleName] = [settingName],
+        };
+
+        Assert.ThrowsException<ArgumentException>(() => GetSettingCommandLineCommand.Execute(requestedSettings));
     }
 }